modify the activity controller to upload the images to amazon

Approved-by: Gabriel Gomez
Approved-by: Jonathan Oscar, Molina Gomez

+++ controllers/activities.js
@@ -1,25 +1,27 @@
 'use-strict'
 
 const { Activity } = require('../models/index');
+const imageServices = require('../services/amazonS3/imageServices')
 
 module.exports = {
 
     postActivity: async (req, res) => {
 
-        const { content, image, name, userId } = req.body;
-
+        const { content, name, userId } = req.body;
+        
         try {
+            const urlOfImage = await imageServices.uploadImage(req.file)
 
             await Activity.create({
                 content,
-                image,
+                image:urlOfImage,
                 name,
                 userId
             });
             return res.status(200).json( { msg: 'Activity posted successfully' } ); // Improve response? 
 
         } catch (error) {
-
+console.log(error.message);
             res.status(400).json(error.message);
 
         } 
@@ -27,14 +29,29 @@ module.exports = {
 
     updateActivity: async (req, res) => {
 
+        
+        console.log(req.body)
+        console.log(req.file)
+        
         const id = req.params.id;
-        const { content, image, name, userId } = req.body;
-
+        const { content, name, userId, urlImage } = req.body;
+        const image =req.file
+        
         try {
+            const urlOfImage = async () => {
+                if(image) {
+                    console.log(urlImage);
+                    const url = await imageServices.uploadImage(image)
+                    await imageServices.deleteImage(urlImage)
+                    return url
+                } else {
+                    return urlImage
+                }
+            }
 
             await Activity.update({
                 content,
-                image,
+                image: await urlOfImage(),
                 name ,
                 userId
             }, { where: { id } } );
@@ -46,6 +63,7 @@ module.exports = {
 
         } catch (error) {
 
+            console.log(error.message)
             res.status(400).json(error.message);
 
         } 
@@ -90,12 +108,14 @@ module.exports = {
         const id = req.params.id;
     
         try {
-
-            const deletedActivity = await Activity.destroy({ where: { id } });
-            
-            return ( deletedActivity ) 
-                ? res.status(200).json({ msg: 'Activity deleted successfuly' }) 
-                : res.status(400).json({ error: 'Activity not Found' });
+            const activityToDelete = await Activity.findByPk(id)
+            if (activityToDelete) {
+                await Activity.destroy({ where: { id } });
+                await imageServices.deleteImage(activityToDelete.image)
+                res.status(200).json({ msg: 'Activity deleted successfuly' })
+            }else{
+                res.status(400).json({ error: 'Activity not Found' });
+            }
 
         } catch (error) {
